Move recordingbacktranslationpair to model package
