+++ app/src/main/java/org/sil/storyproducer/model/RecordingBacktranslationsPair.kt
@@ -1,3 +1,3 @@
-package org.sil.storyproducer.controller.keyterm
+package org.sil.storyproducer.model
 
 class RecordingBacktranslationsPair(var recording: String, var backtranslations: ArrayList<String>)
